Change window clearing to whole screen
This is a significant change, instead of trying to clear each widget
from where it was before, we now clear the whole window and then render
each widget, irregardless of whether it is dirty or not.

It's more work, true, but it's almost impossible to continue this way,
given that collection widgets can change size.

With this we can also remove the invalidate call, which makes widget
development easier. so overral it's probably a win.